// src/generated/openapi/models/BenchmarkRunRawDTO.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * VIPLEV API
 * REST API for the VIPLEV benchmarking and test platform
 *
 * The version of the OpenAPI document: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RawTimeSeriesDTO } from './RawTimeSeriesDTO';
import {
    RawTimeSeriesDTOFromJSON,
    RawTimeSeriesDTOFromJSONTyped,
    RawTimeSeriesDTOToJSON,
    RawTimeSeriesDTOToJSONTyped,
} from './RawTimeSeriesDTO';

/**
 * Top-level wrapper for time-series data for a benchmark run, used for XY-chart visualisation
 * @export
 * @interface BenchmarkRunRawDTO
 */
export interface BenchmarkRunRawDTO {
    /**
     * 
     * @type {RawTimeSeriesDTO}
     * @memberof BenchmarkRunRawDTO
     */
    timeSeries?: RawTimeSeriesDTO;
}

/**
 * Check if a given object implements the BenchmarkRunRawDTO interface.
 */
export function instanceOfBenchmarkRunRawDTO(value: object): value is BenchmarkRunRawDTO {
    return true;
}

export function BenchmarkRunRawDTOFromJSON(json: any): BenchmarkRunRawDTO {
    return BenchmarkRunRawDTOFromJSONTyped(json, false);
}

export function BenchmarkRunRawDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): BenchmarkRunRawDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'timeSeries': json['timeSeries'] == null ? undefined : RawTimeSeriesDTOFromJSON(json['timeSeries']),
    };
}

export function BenchmarkRunRawDTOToJSON(json: any): BenchmarkRunRawDTO {
    return BenchmarkRunRawDTOToJSONTyped(json, false);
}

export function BenchmarkRunRawDTOToJSONTyped(value?: BenchmarkRunRawDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'timeSeries': RawTimeSeriesDTOToJSON(value['timeSeries']),
    };
}