// src/generated/openapi/models/DerivedHostSummaryDTO.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * VIPLEV API
 * REST API for the VIPLEV benchmarking and test platform
 *
 * The version of the OpenAPI document: 3.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DerivedServiceSummaryDTO } from './DerivedServiceSummaryDTO';
import {
    DerivedServiceSummaryDTOFromJSON,
    DerivedServiceSummaryDTOFromJSONTyped,
    DerivedServiceSummaryDTOToJSON,
    DerivedServiceSummaryDTOToJSONTyped,
} from './DerivedServiceSummaryDTO';
import type { DerivedResourceSummaryDTO } from './DerivedResourceSummaryDTO';
import {
    DerivedResourceSummaryDTOFromJSON,
    DerivedResourceSummaryDTOFromJSONTyped,
    DerivedResourceSummaryDTOToJSON,
    DerivedResourceSummaryDTOToJSONTyped,
} from './DerivedResourceSummaryDTO';

/**
 * Aggregated resource stats for a single host, with nested service summaries
 * @export
 * @interface DerivedHostSummaryDTO
 */
export interface DerivedHostSummaryDTO {
    /**
     * 
     * @type {string}
     * @memberof DerivedHostSummaryDTO
     */
    hostId?: string;
    /**
     * 
     * @type {string}
     * @memberof DerivedHostSummaryDTO
     */
    hostName?: string;
    /**
     * 
     * @type {DerivedResourceSummaryDTO}
     * @memberof DerivedHostSummaryDTO
     */
    resource?: DerivedResourceSummaryDTO;
    /**
     * 
     * @type {Array<DerivedServiceSummaryDTO>}
     * @memberof DerivedHostSummaryDTO
     */
    services?: Array<DerivedServiceSummaryDTO>;
}

/**
 * Check if a given object implements the DerivedHostSummaryDTO interface.
 */
export function instanceOfDerivedHostSummaryDTO(value: object): value is DerivedHostSummaryDTO {
    return true;
}

export function DerivedHostSummaryDTOFromJSON(json: any): DerivedHostSummaryDTO {
    return DerivedHostSummaryDTOFromJSONTyped(json, false);
}

export function DerivedHostSummaryDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): DerivedHostSummaryDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'hostId': json['hostId'] == null ? undefined : json['hostId'],
        'hostName': json['hostName'] == null ? undefined : json['hostName'],
        'resource': json['resource'] == null ? undefined : DerivedResourceSummaryDTOFromJSON(json['resource']),
        'services': json['services'] == null ? undefined : ((json['services'] as Array<any>).map(DerivedServiceSummaryDTOFromJSON)),
    };
}

export function DerivedHostSummaryDTOToJSON(json: any): DerivedHostSummaryDTO {
    return DerivedHostSummaryDTOToJSONTyped(json, false);
}

export function DerivedHostSummaryDTOToJSONTyped(value?: DerivedHostSummaryDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'hostId': value['hostId'],
        'hostName': value['hostName'],
        'resource': DerivedResourceSummaryDTOToJSON(value['resource']),
        'services': value['services'] == null ? undefined : ((value['services'] as Array<any>).map(DerivedServiceSummaryDTOToJSON)),
    };
}